Oh boy, debug time
Using the new 2023 beta (JAVA 17!!!)

Set up the code for the base swerve drive train that is planned to be used.

Redid Custom PID System in Util to allow suppliers,
Added in Custom Limelight, REVLimitSwitch, and Debounced Controller

A simple wrapper class tbh, same thing new paint

The Limelight class manages all the networking tables making it easier to use.

The Debounced Control handles all of the controller inputs, removing clutter from the RobotContainer file, it also had the axis modification pre applied.

All the debounced systems have been updated to the new Trigger Api, with this calling instant commands through buttons presses are a bit more annoying but there is not much that can be done about this at the moment

Create a custom Swerve Module class to work with the new SwerveDriveKinimatics and SwerveDriveOdometry.
The new class gives us complete control over the module and adds in the now required features not yet supported by SDS

DesiredPosition 2.0 (WIP), Janus - The greek god of motion, Janus is the new name going forward for the Desired Position System. Janus is going under a complete rework from the ground up, being built on a better understanding of trajectories and autonomous routines using odometry

The new swerve module system allows us to use the WPILIB trajectories, Janus will attempt to take advantage of the same concept but done differently to allow easier use of the creation of such routines.

Added in a new system for test motors and subsystems, add the SystemTest interface to a subsystem will allow you to define a function that the TestSystem Command can cycle through to verify the functionality of hardware

Constants are there but most need the correct values put in, I put in all the ones I could find at the time.

Dashboard output needs to setup for the swerve modules and the addition of a limit switch sensor possibly for automatic offset alignment.

Autoselector for smartdash was set up

Everything needs testing a will most definitely break, will continue to add the rest of Janus, the smartdash data

Hopefully this will allow us to jump straight into auto routines as soon as we have a robot. Swervebot or Switchblade testing advised.