Janus Debugging
First round of debugging, system works but the core algorithm needs to be completely redone. Does not properly consider the next point being at a different angle. Need to have it break down the velocity into X and Y, it also needs to consider what speed it needs to slow down to to be able to accurately move in the correct direction to the next point.